Uploaded original code for Expript
This is the original code for Expript. This version is very poorly coded. There are certainly many bugs and issues. I do not recommend using this version at all. I suggest using Expript2, which is half decent, but still not as good as, Expript3, the latest, best, and the version I am currently working on.